neo80_trimode: Do not rely on tapping to detect long presses
The QMK LT() macro is defined as

  #define LT(layer, kc) (QK_LAYER_TAP | (((layer)&0xF) << 8) | ((kc)&0xFF))

in quantum/quantum_keycodes.h. Out of a 16-bit keycode, this macro
requires the upper 4 bits to be 0x4 (QK_LAYER_TAP is 0x4000,
QK_LAYER_TAP_MAX is 0x4fff), reserves the next 4 bits for the layer, and
the lower 8 bits for the keycode to represent. This means that LT() will
truncate key codes above 0xff.

Custom keycodes defined in keyboard.json start with QK_KB_0, which is
0x7e00. LT(0, KC_BT1) therefore expands to

  QK_LAYER_TAP | (0 & 0xf) << 8 | (0x7e01 & 0xff)
  QK_LAYER_TAP | 0 | 1
  0x4001

While this works fine when not used with Via, it is bad user experience
within Via, because it cannot know that we treat the keycode 0x4001 as
KC_BT1, and therefore renders it as LT(0, invalid). Users wishing to
change their keymap would have to know to enter the 0x4001-0x4004
keycodes manually.

Unfortunately tapping does not seem to work without the LT() macro just
directly with the custom key codes. Instead, just start a three second
timer when one of KC_BT1, KC_BT2, KC_BT3, or KC_2G4 are pressed, cancel
the timer when they are released before three seconds have elapsed, and
trigger pairing mode in the timer's callback.

Since tapping is no longer needed, disable it in keyboard.json, and
change the keymap to use the custom key codes directly.

Signed-off-by: Clemens Lang <neverpanic@gmail.com>